# architect-career-api-sdk/openapi/acos-api.yaml
openapi: 3.1.0
info:
  title: Architect Career Operating System API
  description: Platform API documentation for acos-platform
  contact:
    name: ACOS Engineering
    email: platform@acos.local
  license:
    name: Proprietary
  version: development
servers:
  - url: http://127.0.0.1:8080
    description: Generated server url
security:
  - bearer-jwt: []
...
paths:
...
  /api/v1/portfolio/achievements/{achievementId}:
    get:
      tags:
        - Portfolio Achievements
      summary: Get achievement
      operationId: get_4
      parameters:
        - name: achievementId
          in: path
          description: Achievement identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAchievementResponse'
      security:
        - bearer-jwt: []
    put:
      tags:
        - Portfolio Achievements
      summary: Update achievement
      operationId: update_4
      parameters:
        - name: achievementId
          in: path
          description: Achievement identifier
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AchievementRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAchievementResponse'
      security:
        - bearer-jwt: []
    delete:
      tags:
        - Portfolio Achievements
      summary: Delete achievement
      operationId: delete_4
      parameters:
        - name: achievementId
          in: path
          description: Achievement identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
      security:
        - bearer-jwt: []
...
  /api/v1/portfolio/achievements:
    get:
      tags:
        - Portfolio Achievements
      summary: List achievements
      operationId: list_4
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseListAchievementResponse'
      security:
        - bearer-jwt: []
    post:
      tags:
        - Portfolio Achievements
      summary: Create achievement
      operationId: create_4
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AchievementRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAchievementResponse'
      security:
        - bearer-jwt: []
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDetail'
...
    FieldErrorDetail:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
        rejectedValue: {}
...
    AchievementRequest:
      type: object
      description: Payload used to create or update an achievement
      properties:
        title:
          type: string
          description: Achievement title
          example: Led platform modernization
          maxLength: 200
          minLength: 0
        description:
          type: string
          description: Achievement description
          example: Migrated monolith to modular Spring services.
          maxLength: 2000
          minLength: 0
        achievedOn:
          type: string
          format: date
          description: Date achieved
          example: '2025-11-15'
        organization:
          type:
            - string
            - 'null'
          description: Optional organization
          example: ACME Corp
          maxLength: 200
          minLength: 0
      required:
        - achievedOn
        - description
        - title
    AchievementResponse:
      type: object
      description: Portfolio achievement
      properties:
        id:
          type: string
          format: uuid
          description: Achievement identifier
        title:
          type: string
          description: Achievement title
        description:
          type: string
          description: Achievement description
        achievedOn:
          type: string
          format: date
          description: Date achieved
        organization:
          type:
            - string
            - 'null'
          description: Optional organization
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
      required:
        - achievedOn
        - createdAt
        - description
        - id
        - title
        - updatedAt
    ApiResponseAchievementResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/AchievementResponse'
        error:
          $ref: '#/components/schemas/ApiError'
        correlationId:
          type: string
        timestamp:
          type: string
          format: date-time
...
    ApiResponseListAchievementResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/AchievementResponse'
        error:
          $ref: '#/components/schemas/ApiError'
        correlationId:
          type: string
        timestamp:
          type: string
          format: date-time
...
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        data: {}
        error:
          $ref: '#/components/schemas/ApiError'
        correlationId:
          type: string
        timestamp:
          type: string
          format: date-time
...
  securitySchemes:
    bearer-jwt:
      type: http
      description: JWT access token issued by POST /api/v1/auth/login
      name: bearer-jwt
      scheme: bearer
      bearerFormat: JWT